feat: Extend PluginContext to include base plugins
